refactor: split commands into multiple modules to reduce #cfg directives
