+++ cli/Cargo.toml
@@ -15,3 +15,4 @@ clap.workspace = true
 flexispot_e7_controller_lib.workspace = true
 flexispot_e7_controller_web.workspace = true
 ureq = { version = "2.9.1", default-features = false, features = ["json"] }
+url = "2.5.0"
